20/04/2015
cambios cierre

+++ new/tables/ad_tcier.sql
@@ -2,9 +2,9 @@ CREATE TABLE AD_TCIER(
     CIER_CIER                   SERIAL                                  ,
     CIER_FECH                   TIMESTAMP       NOT NULL                ,
     CIER_USUA                   INT             NOT NULL                ,
-    CIER_VLRI                   INT             NOT NULL                ,
-    CIER_VLRT                   INT      		NOT NULL			    ,
-    CIER_VLRC                 	INT			    NOT NULL                ,
+    CIER_VLRI                   NUMERIC(15,5)             NOT NULL                ,
+    CIER_VLRT                   NUMERIC(15,5)      		NOT NULL			    ,
+    CIER_VLRC                 	NUMERIC(15,5)			    NOT NULL                ,
     CIER_SEDE                   INT     		NOT NULL                ,
     CIER_ESTADO                 VARCHAR(2)    	NOT NULL                ,
 PRIMARY KEY (CIER_CIER)
